requirements.txt
click==8.1.2
imageio==2.17.0
numpy==1.22.3
Pillow==9.1.0
tqdm==4.64.0